Added debugging functionality and data file
